requirements.txt
streamlit
google-generativeai
python-dotenv
pandas
requests
pyPDF2